Initial code
This is my first version in the creation of the utility and begin with the exploration of GitHub itself and the bash shell-scripting language. The result of this code is the result form one of examples in lab 9. However previously to the example and attending earlier labs and lectures aided my understanding in how the code is executing. Finding the sensitivity of the code difficult to get use to for example leaving a line of white space within the code causes the utility to crash.